defect-inflow-predictor/requirements.txt
pandas
pyyaml
numpy
requests
